+++ ErsatzTV.Application/MediaCollections/Commands/DeleteSmartCollection.cs
@@ -1,0 +1,7 @@
+﻿using ErsatzTV.Core;
+using LanguageExt;
+
+namespace ErsatzTV.Application.MediaCollections.Commands
+{
+    public record DeleteSmartCollection(int SmartCollectionId) : MediatR.IRequest<Either<BaseError, Unit>>;
+}