+++ ErsatzTV.Core/Domain/ConfigElementKey.cs
@@ -18,6 +18,7 @@
         public static ConfigElementKey ChannelsPageSize => new("pages.channels.page_size");
         public static ConfigElementKey CollectionsPageSize => new("pages.collections.page_size");
         public static ConfigElementKey MultiCollectionsPageSize => new("pages.multi_collections.page_size");
+        public static ConfigElementKey SmartCollectionsPageSize => new("pages.smart_collections.page_size");
         public static ConfigElementKey SchedulesPageSize => new("pages.schedules.page_size");
         public static ConfigElementKey SchedulesDetailPageSize => new("pages.schedules.detail_page_size");
         public static ConfigElementKey PlayoutsPageSize => new("pages.playouts.page_size");